requirements/hpc.txt
# HPC-specific dependencies (minimal set for cluster environments)
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0.0
accelerate>=0.20.0

# Additional HPC utilities
wandb>=0.15.0  # For experiment tracking (optional)
tensorboard>=2.13.0  # For logging (optional)
